test: reorganize UTs
