+++ src/morphism.spec.ts
@@ -1,54 +1,7 @@
-import Morphism, { morphism, StrictSchema } from './morphism';
-
-class User {
-  firstName?: string;
-  lastName?: string;
-  phoneNumber?: string;
-  type?: string;
-
-  groups: Array<any> = new Array<any>();
-
-  constructor(firstName?: string, lastName?: string, phoneNumber?: string) {
-    this.firstName = firstName;
-    this.lastName = lastName;
-    this.phoneNumber = phoneNumber;
-
-    this.type = 'User'; // Use to test default value scenario
-  }
-
-  /**
-   * Use to test runtime access to the created object context
-   * @param {} group
-   * @param {} externalTrigger
-   */
-  addToGroup(group: any, externalTrigger: any) {
-    this.groups.push(group);
-    externalTrigger(this, group);
-  }
-}
+import Morphism, { StrictSchema, morphism, Schema } from './morphism';
+import { User, MockData } from './utils-test';
 
 describe('Morphism', () => {
-  interface MockData {
-    firstName: string;
-    lastName: string;
-    age: number;
-    address: {
-      streetAddress: string;
-      city: string;
-      state: string;
-      postalCode: string;
-    };
-    phoneNumber: [
-      {
-        type: string;
-        number: string;
-      },
-      {
-        type: string;
-        number: string;
-      }
-    ];
-  }
   const dataToCrunch: MockData[] = [
     {
       firstName: 'John',
@@ -77,6 +30,71 @@ describe('Morphism', () => {
     Morphism.register(User);
   });
 
+  describe('Currying Function overload', () => {
+    it('Should return a collection of objects when an array is provided as source', () => {
+      const schema = { foo: 'bar' };
+      const res = morphism(schema, [{ bar: 'test' }]);
+      expect(res.map).toBeDefined();
+      expect(res[0].foo).toEqual('test');
+    });
+    it('Should return a single object matching the schema structure when an object is provided as source', () => {
+      const schema = { foo: 'bar' };
+      const res = morphism(schema, { bar: 'test' });
+
+      expect(res.foo).toEqual('test');
+    });
+
+    it('Should return a Mapper which outputs a Class Object when a Class Type is specified and no items', () => {
+      class Foo {
+        foo: string;
+      }
+      const schema = { foo: 'bar' };
+      const source = { bar: 'value' };
+      const mapper = morphism(schema, null, Foo);
+      expect(mapper(source).foo).toEqual('value');
+      expect(mapper([source][0]).foo).toEqual('value');
+    });
+
+    it('Should return a Mapper which outputs a Typed Object from the generic provided', () => {
+      interface IFoo {
+        foo: string;
+      }
+      const schema: Schema<IFoo> = { foo: 'bar' };
+      const source = { bar: 'value' };
+      const mapper = morphism(schema);
+
+      expect(mapper(source).foo).toEqual('value');
+      expect(mapper([source][0]).foo).toEqual('value');
+    });
+
+    it('Should do a straight mapping with an Interface provided', () => {
+      interface Destination {
+        foo: string;
+        bar: string;
+        qux: string;
+      }
+      interface Source {
+        bar: string;
+      }
+      const schema: StrictSchema<Destination, Source> = {
+        foo: 'bar',
+        bar: 'bar',
+        qux: elem => elem.bar
+      };
+      const source = { bar: 'value' };
+
+      const target = morphism(schema, source);
+      const targets = morphism(schema, [source]);
+      const singleTarget = targets.shift();
+
+      expect(target.foo).toEqual('value');
+      expect(singleTarget).toBeDefined();
+      if (singleTarget) {
+        expect(singleTarget.foo).toEqual('value');
+      }
+    });
+  });
+
   describe('Plain Objects', function() {
     it('should export Morphism function curried function', function() {
       expect(typeof Morphism).toEqual('function');
@@ -89,12 +107,6 @@ describe('Morphism', () => {
 
     it('should provide an Object as result when Morphism is applied on an Object', function() {
       expect(Morphism({}, {})).toEqual({});
-    });
-
-    it('should throw an exception when setting a mapper with a falsy schema', function() {
-      expect(() => {
-        Morphism.setMapper(User, null as any);
-      }).toThrow();
     });
 
     it('should throw an exception when trying to access a path from an undefined object', function() {
@@ -402,299 +414,6 @@ describe('Morphism', () => {
       };
       let results = Morphism(schema, dataToCrunch);
       expect(results[0]).toEqual(desiredResult);
-    });
-  });
-
-  describe('Mappers Registry', function() {
-    it('should throw an exception when using Registration function without parameters', function() {
-      expect(() => Morphism.register(null as any, null)).toThrow();
-    });
-
-    it('should throw an exception when trying to register a mapper type more than once', function() {
-      expect(() => {
-        Morphism.register(User, {});
-      }).toThrow();
-    });
-
-    it('should return the stored mapper after a registration', function() {
-      let schema = {
-        phoneNumber: 'phoneNumber[0].number'
-      };
-      let mapper = Morphism.setMapper(User, schema);
-      let mapperSaved = Morphism.getMapper(User);
-      expect(typeof mapper).toEqual('function');
-      expect(typeof mapperSaved).toEqual('function');
-      expect(mapperSaved).toEqual(mapper);
-    });
-
-    it('should get a stored mapper after a registration', function() {
-      Morphism.setMapper(User, {});
-      expect(typeof Morphism.getMapper(User)).toEqual('function');
-    });
-
-    it('should allow to map data using a registered mapper', function() {
-      let schema = {
-        phoneNumber: 'phoneNumber[0].number'
-      };
-      Morphism.setMapper(User, schema);
-      let desiredResult = new User('John', 'Smith', '212 555-1234');
-      expect(Morphism.map(User, dataToCrunch)).toBeTruthy();
-      expect(Morphism.map(User, dataToCrunch)[0]).toEqual(desiredResult);
-    });
-
-    it('should allow to map data using a mapper updated schema', function() {
-      let schema = {
-        phoneNumber: 'phoneNumber[0].number'
-      };
-      let mapper = Morphism.setMapper(User, schema);
-      let desiredResult = new User('John', 'Smith', '212 555-1234');
-      expect(mapper(dataToCrunch)[0]).toEqual(desiredResult);
-    });
-
-    it('should throw an exception when trying to set an non-registered type', function() {
-      Morphism.deleteMapper(User);
-      expect(() => {
-        Morphism.setMapper(User, {});
-      }).toThrow();
-    });
-  });
-
-  describe('Class Type Mapping', function() {
-    beforeEach(() => {
-      Morphism.deleteMapper(User);
-    });
-
-    it('should use the constructor default value if source value is undefined', function() {
-      let sourceData: any = {
-        firstName: 'John',
-        lastName: 'Smith',
-        type: undefined // <== this field should fallback to the type constructor default value
-      };
-      let desiredResult = new User('John', 'Smith');
-      let mapper = Morphism.register(User);
-      expect(desiredResult.type).toEqual('User');
-      expect(mapper([sourceData])[0]).toEqual(desiredResult);
-    });
-
-    it('should allow straight mapping from a Type without a schema', () => {
-      let userName = 'user-name';
-      let user = Morphism(null as any, { firstName: userName }, User);
-      expect(user).toEqual(new User(userName));
-    });
-
-    it('should allow straight mapping from a Type with a schema', () => {
-      let dataSource = {
-        userName: 'a-user-name'
-      };
-      let schema = {
-        firstName: 'userName'
-      };
-      let user = Morphism(schema, dataSource, User);
-      expect(user).toEqual(new User(dataSource.userName));
-    });
-
-    it('should pass created object context for complex interractions within object', function() {
-      let dataSource = {
-        groups: ['main', 'test']
-      };
-
-      let triggered = false;
-      let trigger = (_user: User, _group: any) => {
-        triggered = true;
-      };
-
-      let schema = {
-        groups: (object: any, _items: any, constructed: User) => {
-          if (object.groups) {
-            for (let group of object.groups) {
-              constructed.addToGroup(group, trigger);
-            }
-          }
-        }
-      };
-      let user = Morphism(schema, dataSource, User);
-
-      let expectedUser = new User();
-      expectedUser.groups = dataSource.groups;
-      expect(user).toEqual(expectedUser);
-      expect(user.firstName).toEqual(expectedUser.firstName);
-
-      expect(triggered).toEqual(true);
-    });
-
-    it('should return undefined if undefined is given to map without doing any processing', function() {
-      Morphism.register<User, any>(User, { a: 'firstName' });
-      expect(Morphism.map(User, undefined)).toEqual(undefined);
-    });
-
-    it('should override the default value if source value is defined', function() {
-      let sourceData = {
-        phoneNumber: null
-      };
-
-      let mapper = Morphism.register(User, {});
-
-      let result = mapper([sourceData])[0];
-      expect(new User().phoneNumber).toEqual(undefined);
-      expect(result.phoneNumber).toEqual(null);
-    });
-
-    it('should provide an Object as result when Morphism is applied on a typed Object', function() {
-      let mock = {
-        number: '12345'
-      };
-
-      let mapper = Morphism.register(User, { phoneNumber: 'number' });
-      let result = mapper(mock);
-      expect(result.phoneNumber).toEqual(mock.number);
-      expect(result instanceof User).toEqual(true);
-    });
-
-    it('should provide an Object as result when Morphism is applied on a typed Object usin .map', function() {
-      let mock = {
-        number: '12345'
-      };
-
-      Morphism.register(User, { phoneNumber: 'number' });
-      let result = Morphism.map(User, mock);
-      expect(result.phoneNumber).toEqual(mock.number);
-      expect(result instanceof User).toEqual(true);
-    });
-
-    it('should provide a List of Objects as result when Morphism is applied on a list', function() {
-      let mock = {
-        number: '12345'
-      };
-
-      Morphism.register(User, { phoneNumber: 'number' });
-      let result = Morphism.map(User, [mock]);
-      expect(result[0].phoneNumber).toEqual(mock.number);
-      expect(result[0] instanceof User).toBe(true);
-    });
-
-    it('should fallback to constructor default value and ignore function when path value is undefined', function() {
-      let mock = {
-        lastname: 'user-lastname'
-      };
-      let schema = {
-        type: {
-          path: 'unreachable.path',
-          fn: (value: any) => value
-        }
-      };
-
-      Morphism.register(User, schema);
-      expect(new User().type).toEqual('User');
-
-      let result = Morphism.map(User, mock);
-      expect(result.type).toEqual('User');
-    });
-
-    describe('Projection', () => {
-      it('should allow to map property one to one when using `Morphism.map(Type,object)` without registration', function() {
-        let mock = { field: 'value' };
-        class Target {
-          field: any;
-          constructor(field: any) {
-            this.field = field;
-          }
-        }
-        const result = Morphism.map(Target, mock);
-        expect(result).toEqual(new Target('value'));
-      });
-
-      it('should allow to map property one to one when using `Morphism.map(Type,data)` without registration', function() {
-        let mocks = [{ field: 'value' }, { field: 'value' }, { field: 'value' }];
-        class Target {
-          field: any;
-          constructor(field: any) {
-            this.field = field;
-          }
-        }
-        const results = Morphism.map(Target, mocks);
-        results.forEach((res: any) => {
-          expect(res).toEqual(new Target('value'));
-        });
-      });
-
-      it('should allow to use Morphism.map as an iteratee first function', function() {
-        let mocks = [{ field: 'value' }, { field: 'value' }, { field: 'value' }];
-        class Target {
-          field: any;
-          constructor(field: any) {
-            this.field = field;
-          }
-        }
-        const results = mocks.map(Morphism.map(Target));
-        results.forEach(res => {
-          expect(res).toEqual(new Target('value'));
-        });
-      });
-
-      it('should allow to use mapper from `Morphism.map(Type, undefined)` as an iteratee first function', function() {
-        let mocks = [{ field: 'value' }, { field: 'value' }, { field: 'value' }];
-        class Target {
-          field: any;
-          constructor(field: any) {
-            this.field = field;
-          }
-        }
-        const mapper = Morphism.map(Target);
-        const results = mocks.map(mapper);
-        results.forEach(res => {
-          expect(res).toEqual(new Target('value'));
-          expect(res.field).toBeDefined();
-        });
-      });
-
-      it('should accept deep nested actions', () => {
-        interface Source {
-          keyA: string;
-        }
-        const sample: Source = {
-          keyA: 'value'
-        };
-
-        interface Target {
-          keyA: { keyA1: string };
-        }
-
-        const schema: StrictSchema<Target, Source> = { keyA: { keyA1: source => source.keyA } };
-
-        const target = morphism(schema, sample);
-        expect(target).toEqual({ keyA: { keyA1: 'value' } });
-      });
-
-      it('should accept deep nested actions into array', () => {
-        interface Source {
-          keySource: string;
-        }
-        const sample: Source = {
-          keySource: 'value'
-        };
-
-        interface Target {
-          keyA: {
-            keyA1: [
-              {
-                keyA11: string;
-                keyA12: number;
-              }
-            ];
-            keyA2: string;
-          };
-        }
-        const schema: StrictSchema<Target, Source> = {
-          keyA: {
-            keyA1: [{ keyA11: 'keySource', keyA12: 'keySource' }],
-            keyA2: 'keySource'
-          }
-        };
-
-        const target = morphism(schema, sample);
-
-        expect(target).toEqual({ keyA: { keyA1: [{ keyA11: 'value', keyA12: 'value' }], keyA2: 'value' } });
-      });
     });
   });
 });
