+++ src/typescript.spec.ts
@@ -1,7 +1,46 @@
-import { morphism, StrictSchema, Schema } from './morphism';
+import Morphism, { morphism, StrictSchema, Schema } from './morphism';
 
-describe('Morphism', () => {
-  describe('Typescript', () => {
+describe('Typescript', () => {
+  describe('Registry Type Checking', () => {
+    it('Should return a Mapper when using Register', () => {
+      class Foo {
+        foo: string;
+      }
+      const schema = { foo: 'bar' };
+      const source = { bar: 'value' };
+      const mapper = Morphism.register(Foo, schema);
+
+      expect(mapper(source).foo).toEqual('value');
+      expect(mapper([source][0]).foo).toEqual('value');
+    });
+  });
+
+  describe('Schema Type Checking', () => {
+    it('Should allow to type the Schema', () => {
+      interface IFoo {
+        foo: string;
+        bar: number;
+      }
+      const schema: Schema<IFoo> = { foo: 'qux' };
+      const source = { qux: 'foo' };
+      const target = morphism(schema, source);
+
+      expect(target.foo).toEqual(source.qux);
+    });
+
+    it('Should allow to use a strict Schema', () => {
+      interface IFoo {
+        foo: string;
+        bar: number;
+      }
+      const schema: StrictSchema<IFoo> = { foo: 'qux', bar: () => 1 };
+      const source = { qux: 'foo' };
+      const target = morphism(schema, source);
+
+      expect(target.foo).toEqual(source.qux);
+      expect(target.bar).toEqual(1);
+    });
+
     it('should accept 2 generic parameters on StrictSchema', () => {
       interface Source {
         inputA: string;
@@ -67,16 +106,13 @@ describe('Morphism', () => {
       morphism(schema, [{}]);
     });
 
-    xit('should fail with typescript', () => {
+    it('should not fail with typescript', () => {
       interface S {
         s1: string;
       }
       interface D {
         d1: string;
       }
-      const schema: StrictSchema<D, S> = {
-        d1: 's1'
-      };
 
       interface Source {
         boring_api_field: number;
@@ -115,9 +151,7 @@ describe('Morphism', () => {
       if (itemE) {
         itemE.namingIsHard;
       }
-    });
 
-    it('', () => {
       interface S {
         _a: string;
       }
@@ -127,8 +161,6 @@ describe('Morphism', () => {
 
       morphism<D, S>({ a: ({ _a }) => _a.toString() });
       morphism<D, S>({ a: ({ _a }) => _a.toString() });
-      // morphism({ a: ({ b }) => b }, { _a: 'value' });
-      // morphism({ a: ({ b }) => b });
     });
   });
 });
